+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This changelog keeps track of all changes to the Packs SDK. We follow convention
 ### Added
 
 - Added `updateOptions.extraOAuthScopes` to sync tables to support incremental OAuth with 2-way sync.
+- Added `width`, `height`, and `shapeStyle` to `ImageSchema`
 
 ## [1.7.1] - 2023-11-15